Antibiotic:  Antibiotic pharmacodynamic effects are active
Drug "Calculate PD" function tracks antibiotic activity when
there is both an infection present and an antibiotic substance.
The effect is determined by the PD index of the substance (e.g.
time above MIC) in relation to the minimun inhibitory concentration
defined for the infection.  This effect is used to scale down
the bacteria growth rate in the inflammation model.